add driver parameter caching
