+++ client_integration_test.go
@@ -36,7 +36,7 @@ func TestGetEnginePropsByName(t *testing.T) {
 
 // TestQuery tests simple query
 func TestQuery(t *testing.T) {
-	queryResponse, err := clientMock.Query(context.TODO(), engineUrlMock, "SELECT 1", map[string]string{"database": databaseMock})
+	queryResponse, err := clientMock.Query(context.TODO(), engineUrlMock, "SELECT 1", map[string]string{"database": databaseMock}, func(key, value string) {})
 	if err != nil {
 		t.Errorf("Query returned an error: %v", err)
 	}
@@ -52,10 +52,10 @@ func TestQuery(t *testing.T) {
 // TestQuery with set statements
 func TestQuerySetStatements(t *testing.T) {
 	query := "SELECT * FROM information_schema.tables"
-	if _, err := clientMock.Query(context.TODO(), engineUrlMock, query, map[string]string{"use_standard_sql": "1", "database": databaseMock}); err != nil {
+	if _, err := clientMock.Query(context.TODO(), engineUrlMock, query, map[string]string{"use_standard_sql": "1", "database": databaseMock}, func(key, value string) {}); err != nil {
 		t.Errorf("Query returned an error: %v", err)
 	}
-	if _, err := clientMock.Query(context.TODO(), engineUrlMock, query, map[string]string{"use_standard_sql": "0", "database": databaseMock}); err == nil {
+	if _, err := clientMock.Query(context.TODO(), engineUrlMock, query, map[string]string{"use_standard_sql": "0", "database": databaseMock}, func(key, value string) {}); err == nil {
 		t.Errorf("Query didn't return an error, but should")
 	}
 }
