+++ client_v0_test.go
@@ -154,7 +154,7 @@ func TestUserAgentV0(t *testing.T) {
 	client.accessTokenGetter = client.getAccessToken
 	client.parameterGetter = client.getQueryParams
 
-	_, _ = client.Query(context.TODO(), server.URL, "SELECT 1", map[string]string{})
+	_, _ = client.Query(context.TODO(), server.URL, "SELECT 1", map[string]string{}, func(key, value string) {})
 	if userAgentHeader != userAgentValue {
 		t.Errorf("Did not set User-Agent value correctly on a query request")
 	}
@@ -175,7 +175,7 @@ func TestProtocolVersionV0(t *testing.T) {
 	client.accessTokenGetter = client.getAccessToken
 	client.parameterGetter = client.getQueryParams
 
-	_, _ = client.Query(context.TODO(), server.URL, "SELECT 1", map[string]string{})
+	_, _ = client.Query(context.TODO(), server.URL, "SELECT 1", map[string]string{}, func(key, value string) {})
 	if protocolVersionValue != protocolVersion {
 		t.Errorf("Did not set Protocol-Version value correctly on a query request")
 	}
@@ -202,7 +202,9 @@ func TestUpdateParametersV0(t *testing.T) {
 	params := map[string]string{
 		"database": "db",
 	}
-	_, err := client.Query(context.TODO(), server.URL, "SELECT 1", params)
+	_, err := client.Query(context.TODO(), server.URL, "SELECT 1", params, func(key, value string) {
+		params[key] = value
+	})
 	if err != nil {
 		t.Errorf("Error during query execution with update parameters header in response %s", err)
 	}
