+++ connection_test.go
@@ -7,7 +7,7 @@ import (
 // TestConnectionPrepareStatement, tests that prepare statement doesn't result into an error
 func TestConnectionPrepareStatement(t *testing.T) {
 	emptyClient := ClientImplV0{}
-	fireboltConnection := fireboltConnection{&emptyClient, "engine_url", map[string]string{}}
+	fireboltConnection := fireboltConnection{&emptyClient, "engine_url", map[string]string{}, nil}
 
 	queryMock := "SELECT 1"
 	_, err := fireboltConnection.Prepare(queryMock)
@@ -20,7 +20,7 @@ func TestConnectionPrepareStatement(t *testing.T) {
 // and prepare statement on closed connection is not possible
 func TestConnectionClose(t *testing.T) {
 	emptyClient := ClientImplV0{}
-	fireboltConnection := fireboltConnection{&emptyClient, "engine_url", map[string]string{}}
+	fireboltConnection := fireboltConnection{&emptyClient, "engine_url", map[string]string{}, nil}
 	if err := fireboltConnection.Close(); err != nil {
 		t.Errorf("Close failed with an err: %v", err)
 	}
